led: add Sequence LED trigger.
Currently, none of available triggers supports playing with the LED brightness
level.  The sequence trigger provides a way to define custom brightness
sequence.
For example, it is easy to customize the sequence to mock up the rhythm of
human breathing which is a nice sequence to tell the user the system is doing
something.

This trigger is meant to be usable for waiting an event to happen, for example
when the system gets ready.  Those sequences may be used to reflect well known
system status (e.g. idle mode, startup...).

This implementation provides several interfaces:
 - to define the sequence itself:
   * plot: definition of plot points using plot or rawplot files,
           each points defines the brightness level
   * interval: constant time interval between each plot point
 - to control the sequence:
   * repeat: the number of repetition of the whole plot sequence
             0 for an infinite loop
   * control: used to control the sequence trigger
     + "start"/"stop": to start/stop the sequence
     + "reset" to clear the sequence counter and the internal plot point index
     + "pause"/"resume" to pause/resume the sequence

Signed-off-by: Gaël PORTAY <gael.portay@gmail.com>